validation-v4/requirements.txt
flask
Flask-Session
chromadb
sentence-transformers
transformers
openai
python-docx
python-dotenv